Moved traversal functions to helper package
